+++ libs/ui/src/index.ts
@@ -7,4 +7,5 @@ export * from './lib/message-toolbar/message-toolbar.component';
 export * from './lib/encryption-dialog/encryption-dialog.service';
 export * from './lib/encryption-dialog/encryption-dialog.tokens';
 export * from './lib/util/with-busy';
+export * from './lib/util/media-query';
 export * from './lib/util/internal-url';